refactor(map): consolidate view-builder marker part files (#3574)
Slice 7: fold the two view-builder marker part fragments
(init_game_map_view_builder_fleet_markers_part.dart and
init_game_map_view_builder_map_markers_part.dart) into a single cohesive
init_game_map_view_builder_markers_part.dart, reducing builder part-file
sprawl from 4 to 3. Both fragments held only private top-level helpers
(no public classes); the combined file stays well under the 1000
non-comment-line cap. Pure relocation; no behavior change.

Refs #3574
