+++ packages/colonizethis_map/lib/src/view/init_game_map_view_builder.dart
@@ -21,9 +21,8 @@ import '../tile_map_grid.dart';
 import '../tile_map_topology_helpers.dart';
 import '../render/tile_map_visualization_shared.dart';
 
-part 'init_game_map_view_builder_fleet_markers_part.dart';
+part 'init_game_map_view_builder_markers_part.dart';
 part 'init_game_map_view_builder_orchestration_part.dart';
 part 'init_game_map_view_builder_cells_markers_part.dart';
-part 'init_game_map_view_builder_map_markers_part.dart';
 
 final _log = packageLogger();
